Split the app and the core support into separate modules
